+++ Demo/Shared/Image+Demo.swift
@@ -14,7 +14,7 @@ extension Image {
     static let documentation = symbol("doc.text")
     static let menu = symbol("ellipsis.circle")
     static let safari = symbol("safari")
-    static let sidebar = symbol("sidebar.leading")
+    static let sidebarLeading = symbol("sidebar.leading")
     static let textEditor = symbol("doc.richtext")
 
     static func symbol(_ named: String) -> Image {